Add example (How to use it in the browser) (#8)
* Initial draft of an example (parsing a p.g. in the browser)

Just a quick&dirty "get it running" example. Uses a local version of the openeo-js-processgraphs repo that already includes bca9649, and a CDN for ajv that is different than stated in the Readme.

* Add example sources for textareas, style with CSS

And add the charset metadata so the warning in the console goes away ;)

* Validate submitted p.g. and output found errors

* Rename function and clean up code

We actually not only parse, but also validate, so that's the more appropriate name. And if I do getElementById and JSON.parse in the same line once, I can do it twice too O:)

* Clean up JS CDN URLs in example

Switched to the ones listed in the readme. The release 1.0.0-beta.5 isn't out yet, but will be sooner or later, and it will then include bca9649, which is crucial for this example to work.